+++ frontends/web/src/components/guide/guide.tsx
@@ -83,18 +83,6 @@ class Guide extends Component<Props> {
     store.setState({ guideExists: false });
   }
 
-  private getEmailText() {
-    const lang = this.props.i18n.language.toUpperCase();
-    const date = new Date().toISOString().split('T')[0];
-    return encodeURI(`mailto:translations@shiftcrypto.ch?subject=BitBoxApp translation feedback (#${lang} ${date})&body=I would like to report a translation issue in the BitBoxApp.
-AppView: ${window.location.pathname}
-AppLanguage: ${this.props.i18n.language}
-(Please keep this information to help us locate the issue)
-
-Description of the translation issue:
-`);
-  }
-
   public render() {
     const { shown, t, children } = this.props;
     return (
@@ -113,14 +101,9 @@ Description of the translation issue:
             <div className={style.entry}>
               {t('guide.appendix.text')}
               {' '}
-              <A href="https://shiftcrypto.ch/contact">{t('guide.appendix.link')}</A>
+              <A href="https://bitbox.swiss/support/">{t('guide.appendix.link')}</A>
               <br />
               <br />
-              Translation feedback:
-              {' '}
-              <A href={this.getEmailText()}>
-                translations@shiftcrypto.ch
-              </A>
             </div>
           </div>
         </div>