+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-## Unreleased
+## 4.38.0
 - Automatically discover used Bitcoin and Litecoin accounts
 - Improve sats amount readability adding thousands separator
 - Revamp settings page for improved UX and aesthetics
 - Update minimum supported buntu version to 20.04+
-- Add support for Czech Crown (CZK)
+- Add support for Czech Crown (CZK) and add Czech translation
 - Add support for Zloty (PLN)
+- Add advanced setup with skip microSD card or 12-word seed options
 
 ## 4.37.0
 - Bundle BitBox02 firmware version v9.14.0